Update addBorder.ts
Updated the code to do what is written in the Expanded Challenge comments. Also worked on better variable names and separation of concerns.

addBorder shouldn't really be concerned with figuring out the longest string, or doing the calculations for the top and bottom borders.

I'm sure this can be refined further, but am moving on for now.